requirements.txt
fastapi
uvicorn
httpx
pydantic
apscheduler
openai
langchain-openai
langchain-core
python-dotenv
